Factor out rustfmt runner
